requirements.txt
# LLM Framework (Updated to v1.0 - Released Nov 2025)
langchain>=1.0.0
langchain-google-genai>=2.0.8
langgraph>=1.0.0
langchain-core>=1.0.0

# Data Validation
pydantic>=2.10.5
pydantic-settings>=2.7.1

# Market Data APIs
alpaca-py>=0.43.0
yfinance>=0.2.49
finnhub-python>=2.4.19

# Search Tool (Updated - no tiktoken conflict)
tavily-python>=0.5.0

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.2
requests>=2.32.3
httpx>=0.28.1

# Development
pytest>=8.3.4
black>=24.11.0
ruff>=0.8.4

# Optional but recommended
langsmith>=0.2.12
